Rename configs as utils
