+++ pyrightconfig.json
@@ -1,6 +1,6 @@
 {
   "ignore": [".tools"],
-  "include": ["src", "examples"],
+  "include": ["src", "tests", "examples"],
   "reportMissingParameterType": "none",
   "reportMissingTypeStubs": "none",
   "reportUnknownArgumentType": "none",
